+++ src/main/java/com/gitlab/cdagaming/craftpresence/config/gui/AdvancedSettingsGui.java
@@ -208,7 +208,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                                                 CraftPresence.GUIS.drawMultiLineString(StringUtils.splitTextByNewLine(
                                                                         ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.item_messages",
                                                                                 ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                                        CraftPresence.TILE_ENTITIES.generatePlaceholderString(attributeName, CraftPresence.TILE_ENTITIES.getListFromName(attributeName))))
+                                                                                        CraftPresence.TILE_ENTITIES.generatePlaceholderString(attributeName, screenInstance.isDebugMode(), CraftPresence.TILE_ENTITIES.getListFromName(attributeName))))
                                                                 ), screenInstance.getMouseX(), screenInstance.getMouseY(), screenInstance.width, screenInstance.height, screenInstance.getWrapWidth(), screenInstance.getFontRenderer(), true);
                                                             }
                                                     )
@@ -238,7 +238,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                                                     CraftPresence.GUIS.drawMultiLineString(StringUtils.splitTextByNewLine(
                                                                             ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.item_messages",
                                                                                     ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                                            CraftPresence.TILE_ENTITIES.generatePlaceholderString(attributeName, CraftPresence.TILE_ENTITIES.getListFromName(attributeName))))
+                                                                                            CraftPresence.TILE_ENTITIES.generatePlaceholderString(attributeName, screenInstance.isDebugMode(), CraftPresence.TILE_ENTITIES.getListFromName(attributeName))))
                                                                     ), screenInstance.getMouseX(), screenInstance.getMouseY(), screenInstance.width, screenInstance.height, screenInstance.getWrapWidth(), screenInstance.getFontRenderer(), true);
                                                                 }
                                                         )
@@ -263,7 +263,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                         StringUtils.splitTextByNewLine(
                                                 ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.item_messages",
                                                         ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                CraftPresence.ENTITIES.generatePlaceholderString("", CraftPresence.TILE_ENTITIES.getListFromName(""))))
+                                                                CraftPresence.ENTITIES.generatePlaceholderString("", isDebugMode(), CraftPresence.TILE_ENTITIES.getListFromName(""))))
                                         ),
                                         getMouseX(), getMouseY(),
                                         width, height,
@@ -314,7 +314,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                                                 CraftPresence.GUIS.drawMultiLineString(StringUtils.splitTextByNewLine(
                                                                         ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_target_messages",
                                                                                 ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                                        CraftPresence.ENTITIES.generatePlaceholderString(attributeName, CraftPresence.ENTITIES.getListFromName(attributeName))))
+                                                                                        CraftPresence.ENTITIES.generatePlaceholderString(attributeName, screenInstance.isDebugMode(), CraftPresence.ENTITIES.getListFromName(attributeName))))
                                                                 ), screenInstance.getMouseX(), screenInstance.getMouseY(), screenInstance.width, screenInstance.height, screenInstance.getWrapWidth(), screenInstance.getFontRenderer(), true);
                                                             }
                                                     )
@@ -344,7 +344,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                                                     CraftPresence.GUIS.drawMultiLineString(StringUtils.splitTextByNewLine(
                                                                             ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_target_messages",
                                                                                     ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                                            CraftPresence.ENTITIES.generatePlaceholderString(attributeName, CraftPresence.ENTITIES.getListFromName(attributeName))))
+                                                                                            CraftPresence.ENTITIES.generatePlaceholderString(attributeName, screenInstance.isDebugMode(), CraftPresence.ENTITIES.getListFromName(attributeName))))
                                                                     ), screenInstance.getMouseX(), screenInstance.getMouseY(), screenInstance.width, screenInstance.height, screenInstance.getWrapWidth(), screenInstance.getFontRenderer(), true);
                                                                 }
                                                         )
@@ -369,7 +369,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                         StringUtils.splitTextByNewLine(
                                                 ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_target_messages",
                                                         ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                CraftPresence.ENTITIES.generatePlaceholderString(CraftPresence.ENTITIES.CURRENT_TARGET_NAME, CraftPresence.ENTITIES.CURRENT_TARGET_TAGS)))
+                                                                CraftPresence.ENTITIES.generatePlaceholderString(CraftPresence.ENTITIES.CURRENT_TARGET_NAME, isDebugMode(), CraftPresence.ENTITIES.CURRENT_TARGET_TAGS)))
                                         ),
                                         getMouseX(), getMouseY(),
                                         width, height,
@@ -420,7 +420,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                                                 CraftPresence.GUIS.drawMultiLineString(StringUtils.splitTextByNewLine(
                                                                         ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_attacking_messages",
                                                                                 ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                                        CraftPresence.ENTITIES.generatePlaceholderString(attributeName, CraftPresence.ENTITIES.getListFromName(attributeName))))
+                                                                                        CraftPresence.ENTITIES.generatePlaceholderString(attributeName, screenInstance.isDebugMode(), CraftPresence.ENTITIES.getListFromName(attributeName))))
                                                                 ), screenInstance.getMouseX(), screenInstance.getMouseY(), screenInstance.width, screenInstance.height, screenInstance.getWrapWidth(), screenInstance.getFontRenderer(), true);
                                                             }
                                                     )
@@ -450,7 +450,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                                                     CraftPresence.GUIS.drawMultiLineString(StringUtils.splitTextByNewLine(
                                                                             ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_attacking_messages",
                                                                                     ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                                            CraftPresence.ENTITIES.generatePlaceholderString(attributeName, CraftPresence.ENTITIES.getListFromName(attributeName))))
+                                                                                            CraftPresence.ENTITIES.generatePlaceholderString(attributeName, screenInstance.isDebugMode(), CraftPresence.ENTITIES.getListFromName(attributeName))))
                                                                     ), screenInstance.getMouseX(), screenInstance.getMouseY(), screenInstance.width, screenInstance.height, screenInstance.getWrapWidth(), screenInstance.getFontRenderer(), true);
                                                                 }
                                                         )
@@ -475,7 +475,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                         StringUtils.splitTextByNewLine(
                                                 ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_attacking_messages",
                                                         ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                CraftPresence.ENTITIES.generatePlaceholderString(CraftPresence.ENTITIES.CURRENT_ATTACKING_NAME, CraftPresence.ENTITIES.CURRENT_ATTACKING_TAGS)))
+                                                                CraftPresence.ENTITIES.generatePlaceholderString(CraftPresence.ENTITIES.CURRENT_ATTACKING_NAME, isDebugMode(), CraftPresence.ENTITIES.CURRENT_ATTACKING_TAGS)))
                                         ),
                                         getMouseX(), getMouseY(),
                                         width, height,
@@ -526,7 +526,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                                                 CraftPresence.GUIS.drawMultiLineString(StringUtils.splitTextByNewLine(
                                                                         ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_riding_messages",
                                                                                 ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                                        CraftPresence.ENTITIES.generatePlaceholderString(attributeName, CraftPresence.ENTITIES.getListFromName(attributeName))))
+                                                                                        CraftPresence.ENTITIES.generatePlaceholderString(attributeName, screenInstance.isDebugMode(), CraftPresence.ENTITIES.getListFromName(attributeName))))
                                                                 ), screenInstance.getMouseX(), screenInstance.getMouseY(), screenInstance.width, screenInstance.height, screenInstance.getWrapWidth(), screenInstance.getFontRenderer(), true);
                                                             }
                                                     )
@@ -556,7 +556,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                                                     CraftPresence.GUIS.drawMultiLineString(StringUtils.splitTextByNewLine(
                                                                             ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_riding_messages",
                                                                                     ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                                            CraftPresence.ENTITIES.generatePlaceholderString(attributeName, CraftPresence.ENTITIES.getListFromName(attributeName))))
+                                                                                            CraftPresence.ENTITIES.generatePlaceholderString(attributeName, screenInstance.isDebugMode(), CraftPresence.ENTITIES.getListFromName(attributeName))))
                                                                     ), screenInstance.getMouseX(), screenInstance.getMouseY(), screenInstance.width, screenInstance.height, screenInstance.getWrapWidth(), screenInstance.getFontRenderer(), true);
                                                                 }
                                                         )
@@ -581,7 +581,7 @@ public class AdvancedSettingsGui extends ExtendedScreen {
                                         StringUtils.splitTextByNewLine(
                                                 ModUtils.TRANSLATOR.translate("gui.config.comment.advanced.entity_riding_messages",
                                                         ModUtils.TRANSLATOR.translate("gui.config.message.tags",
-                                                                CraftPresence.ENTITIES.generatePlaceholderString(CraftPresence.ENTITIES.CURRENT_RIDING_NAME, CraftPresence.ENTITIES.CURRENT_RIDING_TAGS)))
+                                                                CraftPresence.ENTITIES.generatePlaceholderString(CraftPresence.ENTITIES.CURRENT_RIDING_NAME, isDebugMode(), CraftPresence.ENTITIES.CURRENT_RIDING_TAGS)))
                                         ),
                                         getMouseX(), getMouseY(),
                                         width, height,