+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/entity/EntityUtils.java
@@ -354,17 +354,18 @@ public class EntityUtils {
      * Generates Entity Tag Placeholder String
      *
      * @param name The Entity Name
+     * @param addExtraData Whether to add additional data to the string
      * @param tags A List of the tags associated with the Entity
      * @return The Resulting Entity Tag Placeholder String
      */
-    public String generatePlaceholderString(final String name, final List<String> tags) {
+    public String generatePlaceholderString(final String name, final boolean addExtraData, final List<String> tags) {
         final StringBuilder finalString = new StringBuilder();
         if (!tags.isEmpty()) {
             for (String tagName : tags) {
                 finalString.append("\n - &").append(tagName).append("&");
 
-                if (ModUtils.IS_DEV) {
-                    // If in Debug Mode, also append the Tag's value to the placeholder String
+                if (addExtraData) {
+                    // If specified, also append the Tag's value to the placeholder String
                     final String tagValue =
                             tags.equals(CURRENT_TARGET_TAGS) ? CURRENT_TARGET_TAG.getTag(tagName).toString() :
                                     tags.equals(CURRENT_ATTACKING_TAGS) ? CURRENT_ATTACKING_TAG.getTag(tagName).toString() :