+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/gui/impl/SelectorGui.java
@@ -185,7 +185,10 @@ public class SelectorGui extends ExtendedScreen {
     @Override
     public void postRender() {
         final String searchText = ModUtils.TRANSLATOR.translate("gui.config.message.editor.search");
+        final String extraText = isVerboseMode() ? ModUtils.TRANSLATOR.translate("gui.config.title.selector.extra", itemList.size(), originalList.size()) : "";
+        final String displayText = mainTitle + " " + extraText;
+
         renderString(searchText, (30 - (StringUtils.getStringWidth(searchText) / 2f)), (height - 25), 0xFFFFFF);
-        renderString(mainTitle, (width / 2f) - (StringUtils.getStringWidth(mainTitle) / 2f), 15, 0xFFFFFF);
+        renderString(displayText, (width / 2f) - (StringUtils.getStringWidth(displayText) / 2f), 15, 0xFFFFFF);
     }
 }